Added signature to CostDistanceMapOp that takes a vector of source points (Issue #348)
Added unit test for CostDistanceMapOp and modified some of the supporting test classes to support this test.  Also added a python unit test for costDistance.  This test is incomplete because of an issue running python tests right now that the test runner is called unitest.main with an unrecognized verbosity argument.  May be an environment issue.

Signed-off-by: ericwood73 <ericwood73@gmail.com>